refactor: 유지 시간 및 전환 시간 필드를 Double에서 Int로 변경

+++ gs/src/main/kotlin/com/pluxity/patrol/dto/ScenarioResponse.kt
@@ -22,8 +22,8 @@ data class ScenarioResponse(
         val scenarioSceneId: Long,
         val order: Int,
         val scene: SimpleSceneResponse,
-        val duration: Double?,
-        val transitionTime: Double,
+        val duration: Int?,
+        val transitionTime: Int,
     )
 
     data class SimpleSceneResponse(
